[CHANGE] Change to new event. (legacy and new event).
